cloud_functions/pubsub-to-bigquery/requirements.txt
google-cloud-bigquery==3.5.0